+++ NEWS.md
@@ -1,5 +1,11 @@
 # dials (development version)
 
+## Improvements
+
+* The space-filling design functionality was expanded to include several new types of designs: Audze-Eglais, max/min L1, max/min L2, and uniform. These are all pre-computed designs accessed from the sfd package. 
+
+* A new function is used to access all of the space-filling designs called `grid_space_filling()`. 
+
 # dials 1.2.1
 
 ## New parameters